+++ rust/crd/src/affinity.rs
@@ -19,7 +19,10 @@ pub fn get_affinity(
             .iter()
             .filter_map(|catalog| match &catalog.spec.connector {
                 TrinoCatalogConnector::Hive(hive) => Some(&hive.metastore.config_map),
-                TrinoCatalogConnector::Iceberg(iceberg) => Some(&iceberg.metastore.config_map),
+                TrinoCatalogConnector::Iceberg(iceberg) => Some(&iceberg.hive.metastore.config_map),
+                TrinoCatalogConnector::DeltaLake(delta_lake) => {
+                    Some(&delta_lake.hive.metastore.config_map)
+                }
                 TrinoCatalogConnector::BlackHole(_)
                 | TrinoCatalogConnector::Generic(_)
                 | TrinoCatalogConnector::GoogleSheet(_)
@@ -42,7 +45,10 @@ pub fn get_affinity(
                     hive.hdfs.as_ref().map(|hdfs| &hdfs.config_map)
                 }
                 TrinoCatalogConnector::Iceberg(iceberg) => {
-                    iceberg.hdfs.as_ref().map(|hdfs| &hdfs.config_map)
+                    iceberg.hive.hdfs.as_ref().map(|hdfs| &hdfs.config_map)
+                }
+                TrinoCatalogConnector::DeltaLake(delta_lake) => {
+                    delta_lake.hive.hdfs.as_ref().map(|hdfs| &hdfs.config_map)
                 }
                 TrinoCatalogConnector::BlackHole(_)
                 | TrinoCatalogConnector::Generic(_)