+++ rust/crd/src/catalog/delta_lake.rs
@@ -4,7 +4,7 @@ use stackable_operator::schemars::{self, JsonSchema};
 
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
 #[serde(rename_all = "camelCase")]
-pub struct IcebergConnector {
+pub struct DeltaLakeConnector {
     /// The Hive connector exposes the same config
     #[serde(flatten)]
     pub hive: HiveConnector,